+++ tests/caps_word/caps_word_combo/test.mk
@@ -17,3 +17,4 @@ CAPS_WORD_ENABLE = yes
 COMBO_ENABLE = yes
 AUTO_SHIFT_ENABLE = yes
 
+INTROSPECTION_KEYMAP_C = test_combos.c
