+++ tests/test_common/keycode_util.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cstdint>
 #include <string>
 
 std::string get_keycode_identifier_or_default(uint16_t keycode);
